+++ src/Facades/LaravelShield.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Webdevartisan\LaravelShield\Facades;
+namespace Artisanbarista\LaravelShield\Facades;
 
 use Illuminate\Support\Facades\Facade;
 
 /**
- * @see \Webdevartisan\LaravelShield\LaravelShield
+ * @see \Artisanbarista\LaravelShield\LaravelShield
  */
 class LaravelShield extends Facade
 {